requirements.txt
scipy==1.5.2
torch==1.7.0a0+7036e91
torchvision==0.8.0a0
opencv_python==3.4.1.15
nibabel==3.2.1
numpy==1.19.1
yacs==0.1.8
matplotlib==3.3.2
scikit_image==0.15.0
Pillow==8.2.0
skimage==0.0
